update project to nextjs w/ typescript
