+++ components/Footer/index.ts
