+++ next.config.js
@@ -1,0 +1,7 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+	reactStrictMode: false,
+	swcMinify: true,
+};
+
+module.exports = nextConfig;
