+++ package.json
@@ -1,32 +1,36 @@
 {
   "name": "screen-guru",
-  "version": "1.1.0",
+  "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@reach/router": "^1.2.1",
-    "@rematch/core": "^1.2.0",
-    "axios": "^0.18.1",
-    "react": "^16.10.0",
-    "react-color": "^2.13.8",
-    "react-confetti": "^4.0.0",
-    "react-dom": "^16.10.0",
-    "react-github-corners": "^0.2.1",
-    "react-redux": "^7.0.0",
-    "react-scripts": "3.2.0",
-    "react-sizeme": "^2.6.8",
-    "redux": "^4.0.1",
-    "styled-components": "^4.4.0"
-  },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+  "dependencies": {
+    "@rematch/core": "^2.2.0",
+    "axios": "^1.1.3",
+    "next": "13.0.0",
+    "next-head-seo": "^0.1.3",
+    "react": "18.2.0",
+    "react-color": "^2.19.3",
+    "react-confetti": "^6.1.0",
+    "react-dom": "18.2.0",
+    "react-github-corners": "^0.2.1",
+    "react-redux": "^8.0.4",
+    "react-sizeme": "^3.0.2",
+    "redux": "^4.2.0",
+    "styled-components": "^5.3.6"
+  },
+  "devDependencies": {
+    "@types/node": "18.11.8",
+    "@types/react": "18.0.24",
+    "@types/react-color": "^3.0.6",
+    "@types/react-dom": "18.0.8",
+    "@types/styled-components": "^5.1.26",
+    "eslint": "8.26.0",
+    "eslint-config-next": "13.0.0",
+    "typescript": "4.8.4"
+  }
 }
