+++ pages/_app.tsx
@@ -1,0 +1,18 @@
+import type { AppProps } from 'next/app';
+import { init } from '@rematch/core';
+import { Provider } from 'react-redux';
+import models from '../models';
+import 'react-github-corners/dist/GithubCorner.css';
+import '../styles/globals.css';
+
+const store = init({
+	models,
+});
+
+export default function App({ Component, pageProps }: AppProps) {
+	return (
+		<Provider store={store}>
+			<Component {...pageProps} />
+		</Provider>
+	);
+}
