--- src/assets/layout.css
@@ -1,104 +1,0 @@
-body {
-  font-family: 'Lato', sans-serif;
-  height: 100vh;
-  color: #1050ff;
-  margin: 0;
-  padding: 0;
-  border-left: 10px solid #1050ff;
-  border-right: 10px solid #1050ff;
-}
-
-body:before,
-body:after {
-  content: '';
-  position: fixed;
-  background: #1050ff;
-  left: 0;
-  right: 0;
-  height: 10px;
-}
-
-body:before {
-  top: 0;
-}
-
-body:after {
-  bottom: 0;
-}
-
-@media screen and (max-width: 500px) {
-  body {
-    border-left: none;
-    border-right: none;
-  }
-
-  body:before,
-  body:after {
-    height: 0;
-  }
-}
-
-a {
-  color: white;
-}
-
-::selection {
-  color: white;
-  background: #1050ff;
-}
-
-::placeholder {
-  color: #1050ff;
-}
-
-/* https://github.com/daneden/animate.css/blob/master/animate.css */
-
-@keyframes tada {
-  from {
-    -webkit-transform: scale3d(1, 1, 1);
-    transform: scale3d(1, 1, 1);
-  }
-
-  10%,
-  20% {
-    -webkit-transform: scale3d(0.9, 0.9, 0.9) rotate3d(0, 0, 1, -3deg);
-    transform: scale3d(0.9, 0.9, 0.9) rotate3d(0, 0, 1, -3deg);
-  }
-
-  30%,
-  50%,
-  70%,
-  90% {
-    -webkit-transform: scale3d(1.1, 1.1, 1.1) rotate3d(0, 0, 1, 3deg);
-    transform: scale3d(1.1, 1.1, 1.1) rotate3d(0, 0, 1, 3deg);
-  }
-
-  40%,
-  60%,
-  80% {
-    -webkit-transform: scale3d(1.1, 1.1, 1.1) rotate3d(0, 0, 1, -3deg);
-    transform: scale3d(1.1, 1.1, 1.1) rotate3d(0, 0, 1, -3deg);
-  }
-
-  to {
-    -webkit-transform: scale3d(1, 1, 1);
-    transform: scale3d(1, 1, 1);
-  }
-}
-
-.tada {
-  -webkit-animation-name: tada;
-  animation-name: tada;
-}
-
-.animated {
-  -webkit-animation-duration: 1s;
-  animation-duration: 1s;
-  -webkit-animation-fill-mode: both;
-  animation-fill-mode: both;
-}
-
-.animated.infinite {
-  -webkit-animation-iteration-count: infinite;
-  animation-iteration-count: infinite;
-}
